+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 * [breaking] Bunny upgrade fom 0.7.13 to 2.23.0 => changes from single threaded mode to multithreaded mode
 * [breaking] Heartbeats enabled if server offers them
 * [breaking] Publishing timeout enabled and configured by default
+* Add TLS support for redis in deduplication store
 
 ## Version 4.0.0
 * The Beetle::Configuration.server_connection_options now take true and false for the SSL configuration.